requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
Flask-Migrate==4.0.5
python-dateutil==2.8.2
Werkzeug==2.3.7
requests==2.31.0